+++ src/common/pipes/index.ts
@@ -1,1 +1,2 @@
 export * from './validation.pipe';
+export * from './parse-int.pipe';