+++ src/common/pipes/validation.pipe.ts
@@ -1,12 +1,7 @@
-import { HttpException } from '@nestjs/core';
-import {
-	PipeTransform,
-	Pipe,
-	ArgumentMetadata,
-	HttpStatus,
-} from '@nestjs/common';
 import { validate } from 'class-validator';
 import { plainToClass } from 'class-transformer';
+import { PipeTransform } from '../interfaces/pipe-transform.interface';
+import { Pipe, ArgumentMetadata, BadRequestException } from '../index';
 
 @Pipe()
 export class ValidationPipe implements PipeTransform<any> {
@@ -18,9 +13,9 @@ export class ValidationPipe implements PipeTransform<any> {
 		const entity = plainToClass(metatype, value);
 		const errors = await validate(entity);
 		if (errors.length > 0) {
-			throw new HttpException('Validation failed', HttpStatus.BAD_REQUEST);
+			throw new BadRequestException(errors);
 		}
-		return entity;
+		return value;
 	}
 
 	private toValidate(metatype): boolean {