+++ src/common/test/pipes/validation.pipe.spec.ts
@@ -22,20 +22,13 @@ describe('ValidationPipe', () => {
 		target = new ValidationPipe();
 	});
 	describe('transform', () => {
-		describe('when metadata is empty or undefined', () => {
+		describe('when validation passes', () => {
 			it('should return the value unchanged', async () => {
 				const testObj = { prop1: 'value1', prop2: 'value2' };
 				expect(await target.transform(testObj, {} as any)).to.equal(testObj);
 				expect(await target.transform(testObj, {} as any)).to.not.be.instanceOf(
 					TestModel,
 				);
-			});
-		});
-		describe('when metadata contains a class', () => {
-			it('should return an instance of the class', async () => {
-				const testObj = { prop1: 'value1', prop2: 'value2' };
-				const result = await target.transform(testObj, metadata);
-				expect(result).to.be.instanceOf(TestModel);
 			});
 		});
 		describe('when validation vails', () => {