+++ src/index.ts
@@ -6,6 +6,5 @@
  */
 
 export * from './common';
-export * from './core';
 export * from './core/nest-factory';
 export * from './core/nest-application';