+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 4.3.3
+- **common**: `ParseIntPipe` is now available out-of-the-box (`@nestjs/common`)
+- **common**: package contains a set of useful HTTP exceptions now, such as `ForbiddenException`, `UnauthorizedException`, `BadRequestException` etc
+- **core**: `HttpException` was moved to `@nestjs/common`. This one from `core` packages is now DEPRECATED and will be removed in the next MAJOR release
+
 ## 4.3.0
-- **common**: `ValidationPipe` is now available out-of-the-box
+- **common**: `ValidationPipe` is now available out-of-the-box (`@nestjs/common`)
 - **core**: `json` and `urlencoded` (`body-parser`) middlewares are applied by default now, bugfix #252
 - **core** more informative error message (injector) #223
 example: `[ExceptionHandler] Nest can't resolve dependencies of the UsersService (+, +, ?, +, +, +). Please verify whether [2] argument is available in the current context.`